% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_sig_estimate.R
\name{calculate_T_ts}
\alias{calculate_T_ts}
\title{Calculate T tensor from TensorSignature model}
\usage{
calculate_T_ts(tmp_mod)
}
\arguments{
\item{tmp_mod}{TensorSignature model}
}
\description{
Calculate T tensor from TensorSignature model
}
