% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_deviance}
\alias{compute_deviance}
\title{Compute the deviance of a model}
\usage{
compute_deviance(Y, res, return_value = "deviance", X = NULL)
}
\arguments{
\item{Y}{is a count tensor}

\item{res}{is the output of running function{run_EM}.}

\item{X}{is the matrix of covariates.}
}
\value{
deviance
}
\description{
Compute the deviance of a model
}
