% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deviance_curve}
\alias{deviance_curve}
\title{Compute the deviance curve to estimate the number of clusters}
\usage{
deviance_curve(num_sigs = c(), Y, X)
}
\arguments{
\item{num_sigs}{: vector of the number of signatures to use}

\item{Y}{: count tensor}

\item{X}{: matrix of covariates}
}
\value{
data frame with the number of clusters and the deviance
}
\description{
Compute the deviance curve to estimate the number of clusters
}
