% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{estimateEffect}
\alias{estimateEffect}
\title{Compute the estimated effect of the model covariates on signature contribution}
\usage{
estimateEffect(mod0, X, niter, sig_number, to_plot = TRUE)
}
\arguments{
\item{mod0}{: fitted model from STRAND}

\item{niter}{: number of simulations to draw for the estimation of the effect (default = 1000)}

\item{X:}{design matrix with the covariates you wish to estimate the effect of}

\item{sig_number:}{number of the signature you want to estimate the effect of (i.e. 1, or 2, etc.)}

\item{to_plot:}{boolean indicating whether you want to plot the effect (default = TRUE)}
}
\value{
a list containing the summary of the coefficients including p-values, simulated draws, and ggplot object.
}
\description{
Compute the estimated effect of the model covariates on signature contribution
}
