% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_sig_estimate.R
\name{fit_TS}
\alias{fit_TS}
\title{Fit TensorSignature model}
\usage{
fit_TS(Y, X, K, anno_dims, lr = 0.01, tau = 50, max_iter = 1000)
}
\arguments{
\item{Y}{count tensor}

\item{X}{design matrix}

\item{K}{number of signatures}

\item{anno_dims}{annotation dimensions}

\item{lr}{learning rate (default: 0.01)}

\item{tau}{regularization parameter (default: 50)}

\item{max_iter}{maximum number of iterations (default: 1000)}
}
\description{
Fit TensorSignature model
}
