% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_utils.R
\name{generate_data}
\alias{generate_data}
\title{Generate sample data from the structural topic model framework}
\usage{
generate_data(V, K, D, p, no_covars = FALSE, gamma_mean = 0)
}
\arguments{
\item{V}{number of trinucleotide contents (i.e. 96)}

\item{K}{number of signatures or latent factors (i.e. 5)}

\item{D}{number samples (i.e. 100)}

\item{p}{number of covariates (i.e. 3)}

\item{gamma_mean}{mean of effect sizes of covariates in the odds ratio scale.}
}
\value{
list of parameters and sample data in torch tensor format.
}
\description{
Generate sample data from the structural topic model framework
}
