% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_sig_estimate.R
\name{init_tensorsig}
\alias{init_tensorsig}
\title{Initialize TensorSignature algorithm}
\usage{
init_tensorsig(Y, X, K, max_iter, anno_dims)
}
\arguments{
\item{Y}{count tensor}

\item{X}{design matrix}

\item{K}{Number of clusters}

\item{max_iter}{Number of iterations}

\item{anno_dims}{Annotation dimensions}
}
\description{
Initialize TensorSignature algorithm
}
