% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{make_96plot}
\alias{make_96plot}
\title{Generate plot of trinucleotide contexts}
\usage{
make_96plot(mod0, model_type = "STRAND")
}
\arguments{
\item{mod0}{is the runEM output}

\item{model_type}{is the model type i.e. STRAND or TENSIG}
}
\value{
ggplot object with the 96 trinucleotide contexts
}
\description{
Generate plot of trinucleotide contexts
}
