% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_utils.R
\name{NMFinit}
\alias{NMFinit}
\title{Initialize parameters for variational EM algorithm}
\usage{
NMFinit(Y, X, K, max_iter)
}
\arguments{
\item{Y}{count tensor}

\item{X}{matrix of covariates}

\item{K}{number of signatures or latent factors}

\item{max_iter}{maximum number of iterations}
}
\value{
list of torch tensors for input into variational EM
}
\description{
Initialize parameters for variational EM algorithm
}
