% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{plot_factors}
\alias{plot_factors}
\title{Plots factor matrices}
\usage{
plot_factors(mod0, model_type = "STRAND")
}
\arguments{
\item{mod0}{A fitted model object}

\item{model_type}{The model type (e.g. "STRAND")}
}
\value{
matrices with factor terms that can be plotted using pheatmap
}
\description{
Plots factor matrices
}
