% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_utils.R
\name{random_init}
\alias{random_init}
\title{Initialize using random intiial values}
\usage{
random_init(
  Y,
  X,
  anno_dims = list(epi_dim = 16, nuc_dim = 4, clu_dim = 2, V = 96),
  K,
  D
)
}
\arguments{
\item{Y}{count tensor}

\item{X}{tensor of covariates (a p x D matrix)}

\item{anno_dims}{list of dimensions of annotations (default = list(epi_dim = 16, nuc_dim = 4, clu_dim = 2, V= 96))}

\item{K}{number of signatures or latent factors}
}
\description{
Initialize using random intiial values
}
