% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stack}
\alias{stack}
\title{Generate stacked T0 tensor}
\usage{
stack(T0, bt, br, n_epi = 16, n_nuc = 4, n_clu = 2, do_gpu = TRUE)
}
\arguments{
\item{T0}{, tensor}

\item{bt}{, transcriptional bias tensor}

\item{br}{, replication bias tensor}
}
\description{
Generate stacked T0 tensor
}
