% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tf}
\alias{tf}
\title{Function to compute the product of the stacked tensor T0, and F cofactors}
\usage{
tf(T0, covs, missing_rate = NULL)
}
\arguments{
\item{T0}{a tensor}

\item{F}{a list of covariate factors}
}
\value{
T tensor
}
\description{
Function to compute the product of the stacked tensor T0, and F cofactors
}
