% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_sig_estimate.R
\name{transform_E}
\alias{transform_E}
\title{Compute the exposure matrix from a TensorSignature model}
\usage{
transform_E(mod_ts, Y, K, tau = 50, lr = 0.001, max_iter = 10000)
}
\arguments{
\item{mod_ts}{TensorSignature model}

\item{Y}{count tensor}

\item{K}{number of clusters}

\item{tau}{regularization parameter (default=50)}

\item{lr}{learning rate (default=0.001)}

\item{max_iter}{maximum number of iterations (default=1000)}
}
\description{
Compute the exposure matrix from a TensorSignature model
}
