% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calculate_Chat_ts}
\alias{calculate_Chat_ts}
\title{Compute the estimated count matrix from tenosr signature model}
\usage{
calculate_Chat_ts(mod_ts)
}
\arguments{
\item{mod_ts}{fitted model from ts_fit}
}
\description{
Compute the estimated count matrix from tenosr signature model
}
